整理与BUG修复
