+++ src/ZmqNet/ApiContract/Agebull.ZeroNet.Contract.csproj
@@ -20,16 +20,18 @@
     <PackageLicenseUrl>https://github.com/agebullhu/ZmqNet/blob/master/LICENSE</PackageLicenseUrl>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/agebullhu/ZmqNet</RepositoryUrl>
-    <PackageVersion>1.0.6</PackageVersion>
-    <Version>1.0.6</Version>
+    <PackageVersion>1.0.6.1</PackageVersion>
+    <IncludeSymbols>true</IncludeSymbols>
+    <IncludeSource>true</IncludeSource>
+    <Version>1.0.6.1</Version>
     <Authors>Agebull</Authors>
     <Company>agebull</Company>
     <Description>ZeroNet中调用契约定义</Description>
     <PackageReleaseNotes>ZeroNet中调用契约定义</PackageReleaseNotes>
     <PackageId>Agebull.ZeroNet.Contract</PackageId>
     <Product>Agebull.ZeroNet</Product>
-    <AssemblyVersion>1.0.6</AssemblyVersion>
-    <FileVersion>1.0.6</FileVersion>
+    <AssemblyVersion>1.0.6.1</AssemblyVersion>
+    <FileVersion>1.0.6.1</FileVersion>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -81,7 +83,7 @@
     <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Agebull.EntityModel.Core" Version="2.1.2.3" />
+    <PackageReference Include="Agebull.EntityModel.Core" Version="2.1.3" />
     <PackageReference Include="System.ComponentModel.Composition" Version="4.5.0" />
   </ItemGroup>
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
