+++ src/ZmqNet/ZmqNetCore/ZeroApi/ApiAction.cs
@@ -1,102 +1,73 @@
 using System;
 using System.Reflection;
-using Agebull.Common.Logging;
-using Agebull.ZeroNet.Core;
 using Newtonsoft.Json;
 
 namespace Agebull.ZeroNet.ZeroApi
 {
+
     /// <summary>
-    /// Api方法的信息
-    /// </summary>
-    public class ApiActionInfo
-    {
-        /// <summary>
-        /// 方法名称
-        /// </summary>
-        public string FunctionName { get; set; }
-        /// <summary>
-        /// Api路由名称
-        /// </summary>
-        public string RouteName { get; set; }
-        /// <summary>
-        /// 访问设置
-        /// </summary>
-        public ApiAccessOption AccessOption { get; set; }
-
-        /// <summary>
-        /// 是否有调用参数
-        /// </summary>
-        public bool HaseArgument { get; set; }
-
-        /// <summary>
-        /// 访问设置
-        /// </summary>
-        public Func<IApiResult> Action { get; set; }
-
-        /// <summary>
-        /// 访问设置
-        /// </summary>
-        public Func<IApiArgument, IApiResult> ArgumentAction { get; set; }
-
-        /// <summary>
-        /// 参数类型
-        /// </summary>
-        public Type ArgumenType { get; set; }
-
-    }
-    /// <summary>
-    /// Api站点
+    ///     Api站点
     /// </summary>
     public abstract class ApiAction
     {
         /// <summary>
-        /// Api名称
+        ///     Api名称
         /// </summary>
         public string Name { get; set; }
+
         /// <summary>
-        /// 访问控制
+        ///     访问控制
         /// </summary>
         public ApiAccessOption Access { get; set; }
 
         /// <summary>
-        /// 需要登录
+        ///     需要登录
         /// </summary>
-        public bool NeedLogin => Access.HasSomeFlags(ApiAccessOption.Customer, ApiAccessOption.Business, ApiAccessOption.Employe);
+        public bool NeedLogin =>
+            Access.HasSomeFlags(ApiAccessOption.Customer, ApiAccessOption.Business, ApiAccessOption.Employe);
+
         /// <summary>
-        /// 是否公开接口
+        ///     是否公开接口
         /// </summary>
         public bool IsPublic => Access.HasFlag(ApiAccessOption.Public);
+
         /// <summary>
-        /// 参数类型
+        ///     参数类型
         /// </summary>
         public Type ArgumenType { get; set; }
+
         /// <summary>
-        /// 还原参数
+        ///     还原参数
         /// </summary>
         public abstract bool RestoreArgument(string argument);
 
         /// <summary>
-        /// 参数校验
+        ///     参数校验
         /// </summary>
         /// <param name="message"></param>
         /// <returns></returns>
         public abstract bool Validate(out string message);
 
         /// <summary>
-        /// 执行
+        ///     执行
         /// </summary>
         /// <returns></returns>
         public abstract IApiResult Execute();
     }
+
     /// <summary>
-    /// Api动作
+    ///     Api动作
     /// </summary>
     public sealed class ApiAction<TResult> : ApiAction
         where TResult : IApiResult
     {
         /// <summary>
-        /// 还原参数
+        ///     执行行为
+        /// </summary>
+        public Func<TResult> Action { get; set; }
+
+        /// <summary>
+        ///     还原参数
         /// </summary>
         public override bool RestoreArgument(string argument)
         {
@@ -104,7 +75,7 @@ namespace Agebull.ZeroNet.ZeroApi
         }
 
         /// <summary>
-        /// 参数校验
+        ///     参数校验
         /// </summary>
         /// <param name="message"></param>
         /// <returns></returns>
@@ -115,11 +86,7 @@ namespace Agebull.ZeroNet.ZeroApi
         }
 
         /// <summary>
-        /// 执行行为
-        /// </summary>
-        public Func<TResult> Action { get; set; }
-        /// <summary>
-        /// 执行
+        ///     执行
         /// </summary>
         /// <returns></returns>
         public override IApiResult Execute()
@@ -127,31 +94,38 @@ namespace Agebull.ZeroNet.ZeroApi
             return Action();
         }
     }
+
     /// <summary>
-    /// API标准委托
+    ///     API标准委托
     /// </summary>
     /// <param name="argument"></param>
     /// <returns></returns>
     public delegate IApiResult ApiDelegate1(IApiArgument argument);
 
     /// <summary>
-    /// API标准委托
+    ///     API标准委托
     /// </summary>
     /// <returns></returns>
     public delegate IApiResult ApiDelegate();
 
     /// <summary>
-    /// Api动作
+    ///     Api动作
     /// </summary>
     public sealed class AnyApiAction<TControler> : ApiAction
         where TControler : class, new()
     {
         /// <summary>
-        /// 参数
+        ///     参数
         /// </summary>
         private IApiArgument _argument;
+
         /// <summary>
-        /// 还原参数
+        ///     执行行为
+        /// </summary>
+        public MethodInfo Method { get; set; }
+
+        /// <summary>
+        ///     还原参数
         /// </summary>
         public override bool RestoreArgument(string argument)
         {
@@ -162,7 +136,7 @@ namespace Agebull.ZeroNet.ZeroApi
         }
 
         /// <summary>
-        /// 参数校验
+        ///     参数校验
         /// </summary>
         /// <param name="message"></param>
         /// <returns></returns>
@@ -175,16 +149,13 @@ namespace Agebull.ZeroNet.ZeroApi
                 message = null;
                 return true;
             }
+
             message = "参数不能为空";
             return false;
         }
-        /// <summary>
-        /// 执行行为
-        /// </summary>
-        public MethodInfo Method { get; set; }
 
         /// <summary>
-        /// 执行
+        ///     执行
         /// </summary>
         /// <returns></returns>
         public override IApiResult Execute()
@@ -203,18 +174,24 @@ namespace Agebull.ZeroNet.ZeroApi
     }
 
     /// <summary>
-    /// Api动作
+    ///     Api动作
     /// </summary>
     public sealed class ApiAction<TArgument, TResult> : ApiAction
         where TArgument : class, IApiArgument
         where TResult : IApiResult
     {
         /// <summary>
-        /// 参数
+        ///     参数
         /// </summary>
         private TArgument _argument;
+
         /// <summary>
-        /// 还原参数
+        ///     执行行为
+        /// </summary>
+        public Func<TArgument, TResult> Action { get; set; }
+
+        /// <summary>
+        ///     还原参数
         /// </summary>
         public override bool RestoreArgument(string argument)
         {
@@ -226,7 +203,7 @@ namespace Agebull.ZeroNet.ZeroApi
         }
 
         /// <summary>
-        /// 参数校验
+        ///     参数校验
         /// </summary>
         /// <param name="message"></param>
         /// <returns></returns>
@@ -239,16 +216,13 @@ namespace Agebull.ZeroNet.ZeroApi
                 message = null;
                 return true;
             }
+
             message = "参数不能为空";
             return false;
         }
 
         /// <summary>
-        /// 执行行为
-        /// </summary>
-        public Func<TArgument, TResult> Action { get; set; }
-        /// <summary>
-        /// 执行
+        ///     执行
         /// </summary>
         /// <returns></returns>
         public override IApiResult Execute()
