+++ src/ZmqNet/ZmqNetCore/ZeroStation/ZeroDiscover.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Agebull.ZeroNet.Core;
-using Agebull.ZeroNet.PubSub;
 
 namespace Agebull.ZeroNet.ZeroApi
 {
@@ -17,14 +16,54 @@ namespace Agebull.ZeroNet.ZeroApi
         /// 主调用程序集
         /// </summary>
         public Assembly Assembly;
-        
+
+        /// <summary>
+        ///     Api方法的信息
+        /// </summary>
+        class ApiActionInfo
+        {
+            /// <summary>
+            ///     方法名称
+            /// </summary>
+            public string FunctionName;
+
+            /// <summary>
+            ///     Api路由名称
+            /// </summary>
+            public string RouteName;
+
+            /// <summary>
+            ///     访问设置
+            /// </summary>
+            public ApiAccessOption AccessOption;
+
+            /// <summary>
+            ///     是否有调用参数
+            /// </summary>
+            public bool HaseArgument;
+
+            /// <summary>
+            ///     无参方法
+            /// </summary>
+            public Func<IApiResult> Action;
+
+            /// <summary>
+            ///     有参方法
+            /// </summary>
+            public Func<IApiArgument, IApiResult> ArgumentAction;
+
+            /// <summary>
+            ///     参数类型
+            /// </summary>
+            public Type ArgumenType;
+        }
         /// <summary>
         /// 查找API
         /// </summary>
         public void FindApies()
         {
             var apiItems = new Dictionary<string, ApiActionInfo>(StringComparer.OrdinalIgnoreCase);
-            var types = Assembly.GetTypes().Where( p=>p.IsSubclassOf(typeof(ZeroApiController))).ToArray();
+            var types = Assembly.GetTypes().Where(p => p.IsSubclassOf(typeof(ZeroApiController))).ToArray();
             foreach (var type in types)
             {
                 FindApi(type, apiItems);
@@ -51,10 +90,22 @@ namespace Agebull.ZeroNet.ZeroApi
         /// </summary>
         /// <param name="type"></param>
         /// <param name="apiItems"></param>
-        private void FindApi(Type  type, Dictionary<string, ApiActionInfo> apiItems)
+        private void FindApi(Type type, Dictionary<string, ApiActionInfo> apiItems)
         {
-            var methods = type.GetMethods(BindingFlags.DeclaredOnly | BindingFlags.Instance 
-                                                                    | BindingFlags.Public 
+            string routeHead = null;
+            var attrib = type.GetCustomAttribute<RouteAttribute>();
+            if (attrib != null)
+            {
+                routeHead = attrib.Name;
+            }
+
+            if (string.IsNullOrWhiteSpace(routeHead))
+                routeHead = null;
+            else
+                routeHead = routeHead.Trim(' ', '\t', '\r', '\n', '/') + "/";
+
+            var methods = type.GetMethods(BindingFlags.DeclaredOnly | BindingFlags.Instance
+                                                                    | BindingFlags.Public
                                                                     | BindingFlags.NonPublic);
 
             foreach (var method in methods)
@@ -62,10 +113,14 @@ namespace Agebull.ZeroNet.ZeroApi
                 var route = method.GetCustomAttribute<RouteAttribute>();
                 if (route == null)
                     continue;
+                var name = route.Name == null 
+                    ? $"{routeHead}{method.Name}" 
+                    : $"{routeHead}{route.Name.Trim(' ', '\t', '\r', '\n', '/')}";
+
                 var info = new ApiActionInfo
                 {
                     FunctionName = method.Name,
-                    RouteName = route.Name
+                    RouteName = name
                 };
                 var pars = method.GetParameters();
                 if (method.GetParameters().Length == 0)
@@ -76,8 +131,7 @@ namespace Agebull.ZeroNet.ZeroApi
                 else if (method.GetParameters().Length == 1)
                 {
                     info.HaseArgument = true;
-                    
-                    info.ArgumentAction = TypeExtend.CreateFunc<IApiArgument, IApiResult>(type.GetTypeInfo(), method.Name,pars[0].ParameterType.GetTypeInfo(), method.ReturnType.GetTypeInfo());
+                    info.ArgumentAction = TypeExtend.CreateFunc<IApiArgument, IApiResult>(type.GetTypeInfo(), method.Name, pars[0].ParameterType.GetTypeInfo(), method.ReturnType.GetTypeInfo());
                     info.ArgumenType = pars[0].ParameterType;
                 }
                 else
@@ -108,6 +162,6 @@ namespace Agebull.ZeroNet.ZeroApi
         }
 
         #endregion
-        
+
     }
 }
